+++ src/realm/object-store/sync/sync_manager.cpp
@@ -598,7 +598,7 @@ std::shared_ptr<SyncSession> SyncManager::get_session(std::shared_ptr<DB> db, co
     auto& client = get_sync_client(); // Throws
     auto path = db->get_path();
 
-    std::lock_guard<std::mutex> lock(m_session_mutex);
+    std::unique_lock<std::mutex> lock(m_session_mutex);
     if (auto session = get_existing_session_locked(path)) {
         sync_config.user->register_session(session);
         return session->external_reference();
@@ -610,7 +610,11 @@ std::shared_ptr<SyncSession> SyncManager::get_session(std::shared_ptr<DB> db, co
     // Create the external reference immediately to ensure that the session will become
     // inactive if an exception is thrown in the following code.
     auto external_reference = shared_session->external_reference();
-
+    // unlocking m_session_mutex here prevents a deadlock for synchronous network
+    // transports such as the unit test suite, in the case where the log in request is
+    // denied by the server: Active -> WaitingForAccessToken -> handle_refresh(401
+    // error) -> user.log_out() -> unregister_session (locks m_session_mutex again)
+    lock.unlock();
     sync_config.user->register_session(std::move(shared_session));
 
     return external_reference;